Add localoam service
This commit adds a new bucket type "localoam" that allows for indexing
metadata locally.

The localoam server has 3 endpoints
 - /list
 - /master.json
 - /*.json
`/list` lists the json files at the attached docker volume at `/data`
`/master.json` serves an oin-register like file with a new type
'localoam'
Any other query to the endpoint is returned as an ordinary file server
request with the appropriate LastModifed and content-type headers.

Additionally, the catalog now has a plugin 'localoam' that allows for
indexing this type of bucket.