Integration test setup for uploading/processing
This is a significant change to how the app is setup.

Config has been strictly refactored. Striving for all ENV in .env and
accessing all config through values set in config.js.

I suspect the build process had become unfocussed overtime so I've
attempted to somewhat start from scratch.

There is now a simpler Dockerfile and docker-compose setup and
LocalOAM has been removed. There is now no test-specific code in the
codebase.

Includes a single test that runs against the dockerised API, uploads
some imagery and waits for it to be processed.